condaenv.xj5zwj0d.requirements.txt
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botbuilder-dialogs==4.16.2
botbuilder-ai==4.16.2
botbuilder-azure==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-testing==4.16.2
azure-functions>=1.17.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry==1.6.4
opentelemetry-sdk>=1.20.0
opentelemetry-api>=1.20.0
opentelemetry-instrumentation>=0.49b0
snowflake-snowpark-python>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pytest-mock>=3.10.0
pytest-env>=1.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aiohttp==3.10.5
requests==2.32.3
jsonpickle==1.4.2
pyyaml>=6.0.0
rdflib>=6.0.0
owlready2>=0.37
pyshacl>=0.20.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
myst-parser>=2.0.0